Redraw Heatmap as new scrobbles are being fetched
Previously, Heatmap was only drawn once all scrobbles were
fetched, now uses componentDidMount() to draw the Heatmap frame
(Legend,Days, and Texts), and componentDidUpdate() to update
the fill value for every Day.

Now the days have IDs based on the date, and not in the number
of days elapsed since startDate.

`dateToUTC()` was useless, since it returned a different epoch
for every time zone, and the date on the scrobbles was in UTC.
Now, `Date.getTime()/1000` is used in the Heatmap constructor,
and returns an epoch in UTC.

Now, only the scrobbles that are between `startDate` and
`endDate` are processed, ignoring any date before `startDate`
or after `endDate`.

Three functions were removed:
  - `makeScrobblingFrequencyArray()`, since it was divided in
    two: `componentDidUpdate()` makes a dictionary using a
    generated id from the date, and a default value of 1, and
    incremented every time the same id is encountered in the
    array. Then, `d3.entries()` converts the dictionary
    {id:count, id:count} into an array [{key: id, value: count}],
    which makes it iterable.
  - `getScrobblesInPeriod()`, since the process of deciding what
    scrobbles are inside the period of the Heatmap  is managed by
    considering the starting and ending dates of every API response.
  - `daysBetweenDates()` since it is completely useless by now.
    IDs are generated using Dates, and not Days since Start.

Only thing left is making proper comments.